refactor(client): sign-in service
